Crate odilia_tower

Integrations for the Odilia screen reader and the tower ecosystem. While most Tower crates focus on networking applications, we beleive that it has great usefulness to us as assistive technology makers.

Most of these are generic-ish utilities we’ve needed in Odilia. We are open to removing them in the case that general-purpose ones are avaialbe and easy to integrate.

Re-exports

pub use service_ext::ServiceExt;

Modules

async_try

A trait AsyncTryFrom, its associated layer AsyncTryIntoLayer, and a blanket implementation of odilia_common::from_state::TryFromState.

service_ext

Allow postfix notation for building services from existing ones.

state_svc

A generic state provider service. This clones the state upon every invocation of Service::call, so make sure it’s relatively cheap to do so.

sync_try

Synchronous version of crate::async_try. Create tower::Services out of the ability to convert values via TryFrom and TryInto].

unwrap_svc

A service built specifically to run .unwrap() on computed futures.