odilia_tower

Module unwrap_svc

A service built specifically to run .unwrap() on computed futures.

Why not just call .await.unwrap()? Because when chaining together async functions on stable Rust, if the type is returned from a trait, it must be named (see: unnamable types).

This module also contains a few related “unwrap-then-…” cases that are useful in Odilia.

Structs

FlattenFutResult

A future which flattens a future’s nested results when the outer result in std::convert::Infallible.

MapResponseIntoService

Map the Ok variant of Service::call into a new type. Where the new type R implements From<Res>.

MapResponseTryIntoCommandsService

Maps a response value of Result<T, E> to [<T as TryIntoCommands>::into]

TryIntoCommandFut

Map a future result into return of TryIntoCommands::try_into_commands.

UnwrapFut

A future which unwraps the future’s Future::Output value if it is a Result<T, Infallible> and converts it into T.

Traits

UnwrapFutExt

Add postfix notation for TryIntoCommandFut.