odilia_tower::unwrap_svc

Struct UnwrapFut

pub struct UnwrapFut<F, O, E> {
    fut: F,
    _marker: PhantomData<(O, E)>,
}

A future which unwraps the future’s Future::Output value if it is a Result<T, Infallible> and converts it into T.

This is useful in the context of tower where all services must return Result<T, E>, even if Err(E) will never occur. To ensure safety, this is only possible to use when the E parameter is std::convert::Infallible.

Fields
fut: F_marker: PhantomData<(O, E)>